+++ EatSomewhere/Data/IngredientEntry.cs
@@ -8,7 +8,7 @@ public class IngredientEntry
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public string Id { get; set; }
     public Ingredient Ingredient { get; set; }
-    public double Amount { get; set; }
-    [NotMapped] public int EstimatedCost => (int)double.Round(Ingredient.Cost / Ingredient.Amount * Amount);
+    public decimal Amount { get; set; }
+    [NotMapped] public long EstimatedCost => (long)decimal.Round(Ingredient.Cost / Ingredient.Amount * Amount);
     [JsonIgnore] public List<Food>? Foods { get; set; } = new();
 }
